Refactored duplicate modifier quick fixes into common reusable classes (#878)
* Refactor duplicate modifier quick fixes into common reusable classes

* Changed the return type from ICodeActionId to JakartaCodeActionId

* Changed the return type from ICodeActionId to JakartaCodeActionId

---------
