+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/core/java/codeaction/RemoveStaticModifierQuickFix.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright (c) 2023 IBM Corporation and others.
+* Copyright (c) 2026 IBM Corporation and others.
 *
 * This program and the accompanying materials are made available under the
 * terms of the Eclipse Public License v. 2.0 which is available at
@@ -10,22 +10,20 @@
 * Contributors:
 *     IBM Corporation - initial implementation
 *******************************************************************************/
-package org.eclipse.lsp4jakarta.jdt.internal.persistence;
+package org.eclipse.lsp4jakarta.jdt.core.java.codeaction;
 
-import org.eclipse.lsp4jakarta.commons.codeaction.ICodeActionId;
 import org.eclipse.lsp4jakarta.commons.codeaction.JakartaCodeActionId;
-import org.eclipse.lsp4jakarta.jdt.core.java.codeaction.RemoveModifierConflictQuickFix;
 
 /**
- * Removes the final modifier from the declaring element.
+ * Removes the static modifier from the declaring element.
  */
-public class RemoveFinalModifierQuickFix extends RemoveModifierConflictQuickFix {
+public class RemoveStaticModifierQuickFix extends RemoveModifierConflictQuickFix {
 
     /**
      * Constructor.
      */
-    public RemoveFinalModifierQuickFix() {
-        super("final");
+    public RemoveStaticModifierQuickFix() {
+        super(false, "static");
     }
 
     /**
@@ -33,14 +31,14 @@ public class RemoveFinalModifierQuickFix extends RemoveModifierConflictQuickFix 
      */
     @Override
     public String getParticipantId() {
-        return RemoveFinalModifierQuickFix.class.getName();
+        return RemoveStaticModifierQuickFix.class.getName();
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    protected ICodeActionId getCodeActionId() {
-        return JakartaCodeActionId.PersistenceRemoveFinalModifier;
+    protected JakartaCodeActionId getCodeActionId() {
+        return JakartaCodeActionId.RemoveStaticModifier;
     }
 }
