+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/cdi/RemoveDependentScopeQuickFix.java
@@ -37,7 +37,7 @@ public class RemoveDependentScopeQuickFix extends RemoveAnnotationConflictQuickF
     }
 
     @Override
-    protected ICodeActionId getCodeActionId() {
+    protected JakartaCodeActionId getCodeActionId() {
         return JakartaCodeActionId.CDIRemoveDependentScope;
     }
