+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/cdi/RemoveScopeDeclarationAnnotationsQuickFix.java
@@ -48,7 +48,7 @@ public class RemoveScopeDeclarationAnnotationsQuickFix extends RemoveAnnotationC
      * {@inheritDoc}
      */
     @Override
-    protected ICodeActionId getCodeActionId() {
+    protected JakartaCodeActionId getCodeActionId() {
         return JakartaCodeActionId.CDIRemoveScopeDeclarationAnnotationsButOne;
     }
