+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/cdi/RemoveSingletonAnnotationQuickFix.java
@@ -40,7 +40,7 @@ public class RemoveSingletonAnnotationQuickFix extends RemoveAnnotationConflictQ
      * {@inheritDoc}
      */
     @Override
-    protected ICodeActionId getCodeActionId() {
+    protected JakartaCodeActionId getCodeActionId() {
         return JakartaCodeActionId.CDIRemoveSingletonAnnotation;
     }
 }
