+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/internal/websocket/RemoveWebsocketAnnotationsQuickFix.java
@@ -43,7 +43,7 @@ public class RemoveWebsocketAnnotationsQuickFix extends RemoveAnnotationConflict
     }
 
     @Override
-    protected ICodeActionId getCodeActionId() {
+    protected JakartaCodeActionId getCodeActionId() {
         return JakartaCodeActionId.WBRemoveAnnotation;
     }
